Add support for bridge jobs
