+++ NGitLab.Mock/Clients/PipelineClient.cs
@@ -106,6 +106,16 @@ namespace NGitLab.Mock.Clients
             }
         }
 
+        public IEnumerable<Bridge> GetBridges(PipelineJobQuery query)
+        {
+            throw new NotImplementedException();
+        }
+
+        public GitLabCollectionResponse<Bridge> GetBridgesAsync(PipelineJobQuery query)
+        {
+            throw new NotImplementedException();
+        }
+
         public Models.Job[] GetJobs(int pipelineId)
         {
             using (Context.BeginOperationScope())
