+++ NGitLab/Models/Job.cs
@@ -1,47 +1,14 @@
-﻿using System;
-using System.Runtime.Serialization;
+﻿using System.Runtime.Serialization;
 using System.Text.Json.Serialization;
 
 namespace NGitLab.Models
 {
     [DataContract]
-    public class Job
+    public class Job : JobCommon
     {
-        [DataMember(Name = "name")]
-        [JsonPropertyName("name")]
-        public string Name;
-
-        [DataMember(Name = "id")]
-        [JsonPropertyName("id")]
-        public int Id;
-
-        [DataMember(Name = "ref")]
-        [JsonPropertyName("ref")]
-        public string Ref;
-
-        [DataMember(Name = "commit")]
-        [JsonPropertyName("commit")]
-        public Commit Commit;
-
-        [DataMember(Name = "created_at")]
-        [JsonPropertyName("created_at")]
-        public DateTime CreatedAt;
-
-        [DataMember(Name = "started_at")]
-        [JsonPropertyName("started_at")]
-        public DateTime StartedAt;
-
-        [DataMember(Name = "finished_at")]
-        [JsonPropertyName("finished_at")]
-        public DateTime FinishedAt;
-
-        [DataMember(Name = "stage")]
-        [JsonPropertyName("stage")]
-        public string Stage;
-
-        [DataMember(Name = "coverage")]
-        [JsonPropertyName("coverage")]
-        public double? Coverage;
+        [DataMember(Name = "pipeline")]
+        [JsonPropertyName("pipeline")]
+        public JobPipeline Pipeline;
 
         [DataMember(Name = "artifacts_file")]
         [JsonPropertyName("artifacts_file")]
@@ -51,41 +18,9 @@ namespace NGitLab.Models
         [JsonPropertyName("runner")]
         public JobRunner Runner;
 
-        [DataMember(Name = "pipeline")]
-        [JsonPropertyName("pipeline")]
-        public JobPipeline Pipeline;
-
         [DataMember(Name = "project")]
         [JsonPropertyName("project")]
         public JobProject Project;
-
-        [DataMember(Name = "status")]
-        [JsonPropertyName("status")]
-        public JobStatus Status;
-
-        [DataMember(Name = "tag")]
-        [JsonPropertyName("tag")]
-        public bool Tag;
-
-        [DataMember(Name = "allow_failure")]
-        [JsonPropertyName("allow_failure")]
-        public bool AllowFailure;
-
-        [DataMember(Name = "user")]
-        [JsonPropertyName("user")]
-        public User User;
-
-        [DataMember(Name = "web_url")]
-        [JsonPropertyName("web_url")]
-        public string WebUrl;
-
-        [DataMember(Name = "duration")]
-        [JsonPropertyName("duration")]
-        public float? Duration;
-
-        [DataMember(Name = "queued_duration")]
-        [JsonPropertyName("queued_duration")]
-        public float? QueuedDuration;
 
         [DataContract]
         public class JobRunner
@@ -109,26 +44,6 @@ namespace NGitLab.Models
             [DataMember(Name = "is_shared")]
             [JsonPropertyName("is_shared")]
             public bool IsShared;
-        }
-
-        [DataContract]
-        public class JobPipeline
-        {
-            [DataMember(Name = "id")]
-            [JsonPropertyName("id")]
-            public long Id;
-
-            [DataMember(Name = "ref")]
-            [JsonPropertyName("ref")]
-            public string Ref;
-
-            [DataMember(Name = "sha")]
-            [JsonPropertyName("sha")]
-            public Sha1 Sha;
-
-            [DataMember(Name = "status")]
-            [JsonPropertyName("status")]
-            public JobStatus Status;
         }
 
         [DataContract]
